moving progressIndicator from specific game location to component package of gazeplay-game.commons
