+++ gazeplay-games-commons/src/main/java/net/gazeplay/components/GazeFollowerIndicator.java
@@ -1,4 +1,4 @@
-package net.gazeplay.games.colors;
+package net.gazeplay.components;
 
 import javafx.geometry.Point2D;
 import javafx.scene.input.MouseEvent;
