+++ gazeplay-games-commons/src/main/java/net/gazeplay/components/IGazeProgressIndicator.java
@@ -1,4 +1,4 @@
-package net.gazeplay.games.colors;
+package net.gazeplay.components;
 
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
