+++ gazeplay-games/src/main/java/net/gazeplay/games/colors/ColorsGame.java
@@ -23,6 +23,8 @@ import net.gazeplay.IGameContext;
 import net.gazeplay.commons.configuration.Configuration;
 import net.gazeplay.commons.gaze.devicemanager.GazeEvent;
 import net.gazeplay.commons.ui.Translator;
+import net.gazeplay.components.AbstractGazeIndicator;
+import net.gazeplay.components.GazeFollowerIndicator;
 
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
